Move sleep into goroutine in `LazyDo` (#659)
* Move sleep into goroutine in `LazyDo`

This moves the sleep of up to `threshold` out of the caller's request
path. This allows callers to use `LazyDo` without sleeping for up to
`threshold` when called in succession.

This allows requests to connections that fail to return immediately
instead of waiting for a second when they are selected to lazily perform
the topology refresh.

* move into fn parameters

Signed-off-by: Rueian <rueiancsie@gmail.com>