fix: don't send READONLY for standalone clients (#907)
READONLY is intended for more advanced use cases with Redis Cluster, not
standalone with read replicas. Sending it here appears to be a no-op
that changes the connection flags (from N to r) but otherwise has no
effect.

This also fixes #906, a harmless bug where the primary connection would
unintentionally send READONLY on reconnects due to the shared
ClientOption reference.

Signed-off-by: Rueian <rueiancsie@gmail.com>

# Conflicts:
#	standalone.go